+++ _auto/foldersync-desktop/tools/chocolateyUninstall.ps1
@@ -1,7 +1,7 @@
 ﻿$ErrorActionPreference = 'Stop'
 
 # On Windows 11 24H2, Microsoft broke all *-AppxPackage commands. As a temporary workaround, we use the GAC to install some required assemblies.
-Add-Type -AssemblyName "System.EnterpriseServices"
+Add-Type -AssemblyName 'System.EnterpriseServices'
 $publish = [System.EnterpriseServices.Internal.Publish]::new()
 @(
     'System.Numerics.Vectors.dll',
@@ -13,5 +13,19 @@ $publish = [System.EnterpriseServices.Internal.Publish]::new()
     $publish.GacInstall($dllPath)
 }
 
-$packageFullName = (Get-AppxPackage -AllUsers | Where-Object { $_.Name -eq 'FoldersyncDesktop' }).PackageFullName
-Remove-AppxPackage -Package $packageFullName -AllUsers
+$packageFullNames = (Get-AppxPackage -AllUsers | Where-Object { $_.Name -eq 'FoldersyncDesktop' }).PackageFullName
+foreach ($packageFullName in $packageFullNames) {
+    if (!$packageFullName) {
+        Write-Verbose 'Skipped empty or null PackageFullName.'
+        continue
+    }
+
+    try {
+        Remove-AppxPackage -Package $packageFullName -AllUsers
+    } catch {
+        Write-Warning "Failed to remove package '$packageFullName': $_"
+        if ($_.Exception) {
+            Write-Warning "Exception details: $($_.Exception.Message)"
+        }
+    }
+}
